Fixed optional constant encoding bug; fixed increment integer max value bug
git-svn-id: svn://svn.code.sf.net/p/openfast/code/trunk@138 e6ee02d6-d133-0410-bd08-f9337e8129a4

+++ src/main/java/org/openfast/IntegerValue.java
@@ -45,7 +45,7 @@ public class IntegerValue extends NumericValue {
     }
 
     private boolean equals(NumericValue otherValue) {
-        return value == otherValue.toInt();
+        return value == otherValue.toLong();
     }
     
     public int hashCode() {
